+++ packages/referral_reconciliation/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 1.0.4
 
-* Upgrade digit_ui_components to ^0.0.2-dev.6
+* Updated to latest package versions
 
 ## 1.0.3+1