+++ packages/referral_reconciliation/CHANGELOG.md
@@ -1,6 +1,12 @@
+## 1.0.4+1
+
+* Updated to latest package versions
+
 ## 1.0.4
 
 * Updated to latest package versions
+* UI/UX fixes
+* Bug fixes
 
 ## 1.0.3+1